requirements.txt
pillow==12.1.1
tqdm==4.67.3
